requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
psycopg2==2.8.4
SQLAlchemy==1.3.15
marshmallow==3.5.1
gunicorn==20.0.4
python-dotenv==0.12.0
python-telegram-bot==12.4.2
PyJWT==1.7.1
